+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file. Be aware th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Internal fix for possible taints in event handler
+
 ## [2.6.2] 2025-08-06
 ### Changed
 - Bump TOC to 11.2.0